Fine-tune tests, refactor packages, remove unnecessary message converter
